pyproject.toml
[project]
name = "news"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "deep-translator==1.11.4",
    "dotenv==0.9.9",
    "email==6.0.0a1",
    "google-api-python-client==2.187.0",
    "google-auth==2.45.0",
    "google-auth-oauthlib==1.2.2",
    "langdetect==1.0.9",
    "requests==2.32.5",
]
